// src/teamcity-client/models/properties.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TeamCity REST API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2018.1 (current)
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
// May contain unused imports in some cases
// @ts-ignore
import type { Property } from './property';

/**
 * Represents a list of Property entities.
 * @export
 * @interface Properties
 */
export interface Properties {
  /**
   *
   * @type {number}
   * @memberof Properties
   */
  count?: number;
  /**
   *
   * @type {string}
   * @memberof Properties
   */
  href?: string;
  /**
   *
   * @type {Array<Property>}
   * @memberof Properties
   */
  property?: Array<Property>;
}